add temps

+++ index.php
@@ -11,13 +11,17 @@
         if (!is_dir("music")) {
             mkdir("music");
         }
-        if (isset($_GET['del'])) {
+        if (isset($_GET['del']) && is_file('music/' . $_GET['del'])) {
             $filename = "music/" . htmlspecialchars($_GET['del']);
             unlink($filename);
         }
         if (isset($_POST['name']) && $_POST['name'] != "") {
+            $temp = 200;
+            if (isset($_POST['temps'])) {
+                $temp = htmlspecialchars($_POST['temps']);
+            }
             $i = -1;
-            $content = "";
+            $content = $temp . "/";
             while (isset($_POST['id'.$i])) {
                 $content .= htmlspecialchars($_POST['id'.$i] . "/");
                 $i++;
@@ -27,9 +31,10 @@
             fclose($file);
         }
     ?>
-    <form action="" method="post" id="notes">
+    <form action="index.php" method="post" id="notes">
         <input type="submit" value="Save">
         <input type="text" name='name' placeholder="Name">
+        <input type="text" name='temps' id="temps" placeholder="temps">
     </form>
     <Button id="add">+</Button>
     <Button id="play">Play</Button>
@@ -39,7 +44,7 @@
             $musics = scandir("music");
             foreach ($musics as $music) {
                 if ($music[0] != '.') {
-        ?><form action="" method="get">
+        ?><form action="index.php" method="get">
             <input type="hidden" value="<?php
                 echo file_get_contents("music/".$music);
             ?>">
